% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_check_point}
\alias{euclidean_check_point}
\alias{euclidean_check_tangent}
\title{Check if a point is from (the tangent space of) a Euclidean manifold}
\usage{
euclidean_check_point(x, error = TRUE)

euclidean_check_tangent(v, base_point, error = TRUE)
}
\arguments{
\item{x}{point for which we check if it is on a Euclidean manifold}

\item{error}{a flag to indicate if the function should
throw an error if the check fails}

\item{v, base_point}{tangent vector for we which we check if is in the
tangent space of a Euclidean manifold at \code{base_point}}
}
\value{
\code{TRUE} if the check succeeds, \code{FALSE} if the check fails (
and \code{error = FALSE}).
}
\description{
Check if a point is from (the tangent space of) a Euclidean manifold
}
