% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_log}
\alias{euclidean_log}
\title{Find the tangent vector that connects two points on a rotation manifold}
\usage{
euclidean_log(base_point, target_point)
}
\arguments{
\item{base_point, target_point}{two points on a rotation manifold}
}
\description{
Find the tangent vector that connects two points on a rotation manifold
}
\details{
The inverse of the exponential map on a Euclidean manifold is
\deqn{\log(p, q) = q - p}
where \eqn{p} is the \code{base_point} and \eqn{q} is the \code{target_point}.
}
