% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_map}
\alias{euclidean_map}
\title{Exponential map on a Euclidean manifold}
\usage{
euclidean_map(v, base_point)
}
\arguments{
\item{v}{a tangent vector}

\item{base_point}{a point on a Euclidean manifold}
}
\description{
Go from the \code{base_point} in the direction \code{v}.
}
\details{
The exponential map on a Euclidean manifold is
\deqn{\exp_p(v) = p + v}
where \eqn{p} is the \code{base_point}
}
