% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_random_point}
\alias{euclidean_random_point}
\title{Generate a random point on a Euclidean manifold}
\usage{
euclidean_random_point(n, ...)
}
\arguments{
\item{n}{the dimension of the Euclidean manifold (can be a vector)}

\item{...}{additional parameters passed to \code{\link{rnorm}}#'}
}
\value{
an array
}
\description{
The elements of a Euclidean manifold (also called special orthogonal) are
\eqn{\text{Eu}(n,\cdot) := \{x \in \mathbb{R}^{n\times \cdot}\}}
}
