% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean_random_tangent}
\alias{euclidean_random_tangent}
\title{Generate a random tangent vector to a point on a Euclidean manifold}
\usage{
euclidean_random_tangent(base_point, ...)
}
\arguments{
\item{base_point}{the point from a euclidean manifold}
}
\value{
an array with the same dimensions as \code{base_point}
}
\description{
The elements of the tangent space for a \eqn{p \in \text{SO}(n)}
are skew-symmetric matrices (\eqn{\mathcal{T}_p\text{Eu}(n,\cdot) :=  \{v \in \mathbb{R}^{n\times \cdot}\}})
}
