% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{grassmann_log}
\alias{grassmann_log}
\title{Find the tangent vector that connects two points on a Grassmann manifold}
\usage{
grassmann_log(base_point, target_point)
}
\arguments{
\item{base_point, target_point}{two points on a Grassmann manifold}
}
\description{
Find the tangent vector that connects two points on a Grassmann manifold
}
\details{
The inverse of the exponential map on a Grassmann manifold is
\deqn{\log(p, q) = V \text{diag}(\tan^{-1}(d))U^T}
where \eqn{U\text{diag}(d)V^T = (q^Tp)^{-1} (q - q^Tpp^T)} is a SVD,
\eqn{p} is the \code{base_point} and \eqn{q} is the \code{target_point}.
}
