% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{grassmann_project_point}
\alias{grassmann_project_point}
\alias{grassmann_project_tangent}
\title{Project a point to (the tangent space of ) a Grassmann manifold}
\usage{
grassmann_project_point(x)

grassmann_project_tangent(v, base_point)
}
\arguments{
\item{x, v, base_point}{matrix}
}
\value{
the projected matrix
}
\description{
Project a point to (the tangent space of ) a Grassmann manifold
}
