% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{grassmann_random_point}
\alias{grassmann_random_point}
\title{Generate a random point on a Grassmann manifold}
\usage{
grassmann_random_point(n, k, ...)
}
\arguments{
\item{n, k}{the dimensions of the Grassmann manifold.}
}
\value{
a matrix with \code{n} rows and \code{k} orthogonal columns.
}
\description{
The elements of a Grassmann manifold are represented
by matrices with orthogonal columns
(\eqn{\text{Gr}(n, k) := \{\text{span}(x): x \in \mathbb{R}^{n\times k} | x^T x = I\}}).
}
