% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{grassmann_random_tangent}
\alias{grassmann_random_tangent}
\title{Generate a random tangent vector to a point on a Grassmann manifold}
\usage{
grassmann_random_tangent(base_point, ...)
}
\arguments{
\item{base_point}{the point from a Grassmann manifold. A matrix with orthogonal columns.}

\item{...}{additional parameters passed to \code{\link{rnorm}}}
}
\value{
a matrix with \code{n} rows and \code{k} columns.
}
\description{
The elements of the tangent space for a \eqn{p \in \text{Gr}(n,k)}
are \eqn{\mathcal{T}_p\text{Gr}(n,k) :=  \{v \in \mathbb{R}^n | p^Tv + v^Tp = 0\}}
}
