% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_check_point}
\alias{rotation_check_point}
\alias{rotation_check_tangent}
\title{Check if a point is from (the tangent space of) a Rotation Manifold}
\usage{
rotation_check_point(x, tol = 1e-12, error = TRUE)

rotation_check_tangent(v, base_point, tol = 1e-12, error = TRUE)
}
\arguments{
\item{x}{point for which we check if it is on a rotation manifold}

\item{tol}{the tolerance for the test}

\item{error}{a flag to indicate if the function should
throw an error if the check fails}

\item{v, base_point}{tangent vector for we which we check if is in the
tangent space of a rotation manifold at \code{base_point}}
}
\value{
\code{TRUE} if the check succeeds, \code{FALSE} if the check fails (
and \code{error = FALSE}).
}
\description{
Check if a point is from (the tangent space of) a Rotation Manifold
}
