% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_log}
\alias{rotation_log}
\title{Find the tangent vector that connects two points on a rotation manifold}
\usage{
rotation_log(base_point, target_point)
}
\arguments{
\item{base_point, target_point}{two points on a rotation manifold}
}
\description{
Find the tangent vector that connects two points on a rotation manifold
}
\details{
The inverse of the exponential map on a rotation manifold is
\deqn{\log(p, q) = \log(p^Tq)}
where \eqn{\log(x)} is the matrix logarithm,
\eqn{p} is the \code{base_point} and \eqn{q} is the \code{target_point}.
}
