% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_map}
\alias{rotation_map}
\title{Exponential map on a rotation manifold}
\usage{
rotation_map(v, base_point)
}
\arguments{
\item{v}{a tangent vector}

\item{base_point}{a point on a sphere}
}
\description{
Go from the \code{base_point} in the direction \code{v}.
}
\details{
The exponential map on a rotation manifold is
\deqn{\exp_p(v) = p \exp(v)}
where \eqn{\exp(v)} is the matrix exponential and \eqn{p} is
the \code{base_point}
}
