% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_project_point}
\alias{rotation_project_point}
\alias{rotation_project_tangent}
\title{Project a point to (the tangent space of ) a rotation manifold}
\usage{
rotation_project_point(x)

rotation_project_tangent(v, base_point)
}
\arguments{
\item{x, v, base_point}{square matrix}
}
\value{
the projected matrix
}
\description{
Project a point to (the tangent space of ) a rotation manifold
}
