% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_positive_definite.R
\name{spd_log}
\alias{spd_log}
\title{Find the tangent vector that connects two points on a rotation manifold}
\usage{
spd_log(base_point, target_point)
}
\arguments{
\item{base_point, target_point}{two points on a rotation manifold}
}
\description{
Find the tangent vector that connects two points on a rotation manifold
}
\details{
The inverse of the exponential map on a rotation manifold is
\deqn{\log(p, q) = p^{1/2} \log(p^{-1/2} q p^{-1/2}) p^{1/2}}
where \eqn{\log(x)} is the matrix logarithm,
\eqn{p} is the \code{base_point} and \eqn{q} is the \code{target_point}.
}
