% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_positive_definite.R
\name{spd_map}
\alias{spd_map}
\title{Exponential map on an SPD manifold}
\usage{
spd_map(v, base_point)
}
\arguments{
\item{v}{a tangent vector}

\item{base_point}{a point on a SPD manifold}
}
\description{
Go from the \code{base_point} in the direction \code{v}.
}
\details{
The exponential map on a rotation manifold is
\deqn{\exp_p(v) = p^{1/2} \exp(p^{-1/2} v p^{-1/2}) p^{1/2}}
where \eqn{\exp(v)} is the matrix exponential and \eqn{p} is
the \code{base_point}
}
