% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_positive_definite.R
\name{spd_project_point}
\alias{spd_project_point}
\alias{spd_project_tangent}
\title{Project a point to (the tangent space of ) an SPD manifold}
\usage{
spd_project_point(x, min_eigen_value = 0.01)

spd_project_tangent(v, base_point = NULL)
}
\arguments{
\item{x, v, base_point}{square matrix}

\item{min_eigen_value}{the minimum eigenvalue of the matrix.
Note that making this value too small leads to problems
when calculating the inverse.}
}
\value{
the projected matrix
}
\description{
Project a point to (the tangent space of ) an SPD manifold
}
\references{
Hickham blog post (2021), https://nhigham.com/2021/01/26/what-is-the-nearest-positive-semidefinite-matrix/
}
