% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_positive_definite.R
\name{spd_random_point}
\alias{spd_random_point}
\title{Generate a random point on an SPD manifold}
\usage{
spd_random_point(n, ...)
}
\arguments{
\item{n}{the dimension of the SPD manifold.}
}
\value{
a symmetric positive definite matrix
}
\description{
The elements of an SPD manifold are
\eqn{\text{SPD}(n) := \{x \in \mathbb{R}^{n\times n} | x = x^T,\, a^T x a > 0 \text{ for all } a \in \mathbb{R}^n\}}
}
