% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_positive_definite.R
\name{spd_random_tangent}
\alias{spd_random_tangent}
\title{Generate a random tangent vector to a point on a rotation manifold}
\usage{
spd_random_tangent(base_point, ...)
}
\arguments{
\item{base_point}{the point from a rotation manifold. A square matrix.}
}
\value{
a skew-symmetruc matrix.
}
\description{
The elements of the tangent space for a \eqn{p \in \text{SPD}(n)}
are skew-symmetric matrices (\eqn{\mathcal{T}_p\text{SPD}(n) :=  \{v \in \mathbb{R}^{n\times n} | v = v^T\}})
}
