% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_log}
\alias{sphere_log}
\title{Find the tangent vector that connects two points on a sphere}
\usage{
sphere_log(base_point, target_point)
}
\arguments{
\item{base_point, target_point}{two points on a sphere}
}
\description{
Find the tangent vector that connects two points on a sphere
}
\details{
The inverse of the exponential map on the sphere is
\deqn{\log(p, q) = \frac{\alpha}{\sin(\alpha)}(q - (p^T q) p),}
where \eqn{\alpha = \cos^{-1}(p^T q)}, \eqn{p} is the \code{base_point}
and \eqn{q} is the \code{target_point}.
}
