% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_map}
\alias{sphere_map}
\title{Exponential map on a n-sphere}
\usage{
sphere_map(v, base_point)
}
\arguments{
\item{v}{a tangent vector}

\item{base_point}{a point on a sphere}
}
\description{
Go from the \code{base_point} in the direction \code{v}.
}
\details{
The exponential map on the sphere is
\deqn{\exp_p(v) = \cos(||v||) p + \sin(||v||)\frac{v}{||v||},}
where \eqn{||v||} is the norm of the tangent vector and \eqn{p}
is the \code{base_point}.
}
