% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_project_point}
\alias{sphere_project_point}
\alias{sphere_project_tangent}
\title{Project a point to (the tangent space of ) a sphere}
\usage{
sphere_project_point(x)

sphere_project_tangent(v, base_point)
}
\arguments{
\item{x, v, base_point}{matrix with one column}
}
\value{
the projected matrix
}
\description{
Project a point to (the tangent space of ) a sphere
}
