% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_random_point}
\alias{sphere_random_point}
\title{Generate a random point on the n-sphere}
\usage{
sphere_random_point(n)
}
\arguments{
\item{n}{the dimension of the sphere-manifold. The
familiar sphere in 3D dimensions for \code{n = 2}.}
}
\value{
a matrix with one column and \code{n+1} rows
}
\description{
The elements of a sphere are \eqn{\mathbb{S}^n := \{x \in \mathbb{R}^{n+1} | \sum_i x_i^2 = 1\}}
}
