% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_random_tangent}
\alias{sphere_random_tangent}
\title{Generate a random tangent vector to a point on the n-sphere}
\usage{
sphere_random_tangent(base_point, ...)
}
\arguments{
\item{base_point}{the point from the n-sphere. A matrix with one column.}

\item{...}{additional parameters passed to \code{\link{rnorm}}}
}
\value{
a matrix with one column and \code{nrow(base_point)} rows
}
\description{
The elements of the tangent space for a \eqn{p \in \text{S}(n)}
are \eqn{\mathcal{T}_p\text{S}(n) :=  \{v \in \mathbb{R}^n | \sum_i v_i p_i = 0\}}
}
