% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiefel.R
\name{stiefel_log}
\alias{stiefel_log}
\title{Find the tangent vector that connects two points on a Stiefel manifold}
\usage{
stiefel_log(base_point, target_point, tolerance = 1e-08, max_iter = 100)
}
\arguments{
\item{base_point, target_point}{two points on a sphere}

\item{tolerance, max_iter}{parameters to control the numerical approximation
of the tangent vector calculation.}
}
\description{
Find the tangent vector that connects two points on a Stiefel manifold
}
\details{
For details see algorithm 4 from Zimmermann et al. (2022).
}
\references{
Zimmermann, Ralf, and Knut Hüper. “Computing the Riemannian Logarithm on the Stiefel
Manifold: Metrics, Methods and Performance.” arXiv, February 8, 2022. http://arxiv.org/abs/2103.12046.
}
