% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiefel.R
\name{stiefel_map}
\alias{stiefel_map}
\title{Exponential map on a Stiefel manifold}
\usage{
stiefel_map(v, base_point)
}
\arguments{
\item{v}{a tangent vector}

\item{base_point}{a point on a Stiefel manifold}
}
\description{
Go from the \code{base_point} in the direction \code{v}.
}
\details{
The exponential map on a Stiefel manifold for the canonical
metric. For details please see Edelman et al. (1998).
}
\references{
A. Edelman, T. A. Arias, and S. T. Smith. The geometry of algorithms with orthogonality constraints. SIAM Journal on Matrix Analysis and Applications, 20(2):303–353, 1998.
}
