% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiefel.R
\name{stiefel_project_point}
\alias{stiefel_project_point}
\alias{stiefel_project_tangent}
\title{Project a point to (the tangent space of ) a Stiefel manifold}
\usage{
stiefel_project_point(x)

stiefel_project_tangent(v, base_point)
}
\arguments{
\item{x, v, base_point}{matrix}
}
\value{
the projected matrix
}
\description{
Project a point to (the tangent space of ) a Stiefel manifold
}
\details{
The function uses the \eqn{Q} matrix of a QR-decomposition.
}
