% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiefel.R
\name{stiefel_random_point}
\alias{stiefel_random_point}
\title{Generate a random point on a Stiefel manifold}
\usage{
stiefel_random_point(n, k, ...)
}
\arguments{
\item{n, k}{the dimensions of the Grassmann manifold.}

\item{...}{additional parameters passed to \code{\link{rnorm}}}
}
\value{
a matrix with \code{n} rows and \code{k} orthogonal columns.
}
\description{
The elements of a Grassmann manifold are represented
by matrices with orthogonal columns
(\eqn{\text{St}(n, k) := \{x \in \mathbb{R}^{n\times k} | x^T x = I\}}).
}
