% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R
\name{grassmann_angle_from_tangent}
\alias{grassmann_angle_from_tangent}
\alias{grassmann_angle_from_points}
\title{Find the rotation angle for between elements of Grassmann manifolds}
\usage{
grassmann_angle_from_tangent(v, normalized = TRUE)

grassmann_angle_from_points(base_point, target_point)
}
\arguments{
\item{v}{tanget vector}

\item{normalized}{flag to indicate if the angle is forced to be less than 180 degrees}

\item{base_point, target_point}{two points from a Grassmann manifold}
}
\value{
the angle in degree
}
\description{
Find the rotation angle for between elements of Grassmann manifolds
}
\references{
Bendokat, Thomas, Ralf Zimmermann, and P-A. Absil. "A Grassmann manifold handbook: Basic geometry and computational aspects." arXiv preprint arXiv:2011.13699 (2020).
}
