% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_injective_radius}
\alias{sphere_injective_radius}
\title{The injective radius of a manifold}
\usage{
sphere_injective_radius(n)
}
\arguments{
\item{n}{the dimension of the manifold}
}
\description{
The injective radius of a manifold
}
\details{
The injective radius is the size of the
largest tangent vector for which the exponential map
is a diffeomorphism, i.e., inverting the exponential map
would return the original tangent vector.

On a sphere going from a point \eqn{p} in direction \eqn{v} beyond
its pole to \eqn{q} means that the \eqn{\log(p, q) \ne v}.

Injective radii:
\itemize{
\item Sphere: \eqn{\pi}
}
}
