% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassmann.R, R/rotation.R, R/sphere.R,
%   R/stiefel.R, R/symmetric_positive_definite.R
\name{grassmann_injectivity_radius}
\alias{grassmann_injectivity_radius}
\alias{rotation_injectivity_radius}
\alias{sphere_injectivity_radius}
\alias{stiefel_injectivity_radius}
\alias{spd_injectivity_radius}
\title{The injectivity radius of a manifold}
\usage{
grassmann_injectivity_radius()

rotation_injectivity_radius()

sphere_injectivity_radius()

stiefel_injectivity_radius()

spd_injectivity_radius()
}
\arguments{
\item{n}{the dimension of the manifold}
}
\value{
injectivity radii:
\itemize{
\item Sphere: \eqn{\pi}
\item Rotation: \eqn{\pi/\sqrt{2}}
\item Grassmann: \eqn{\pi/2}
\item Stiefel: \eqn{0.89\,\pi}
\item Symmetric positive definite (SPD): \eqn{\inf}
}
}
\description{
The injectivity radius of a manifold
}
\details{
The injectivity radius is the size of the
largest tangent vector for which the exponential map
is a diffeomorphism, i.e., inverting the exponential map
would return the original tangent vector.

On a sphere going from a point \eqn{p} in direction \eqn{v} beyond
its pole to \eqn{q} means that the \eqn{\log(p, q) \ne v}.
}
\references{
Rentmeesters, Quentin. “Algorithms for Data Fitting on Some Common Homogeneous Spaces,” (2013)
}
