% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_random_tangent}
\alias{sphere_random_tangent}
\title{Generate a random tangent vector to a point on the n-sphere}
\usage{
sphere_random_tangent(base_point, ...)
}
\arguments{
\item{base_point}{the point from the n-sphere. A matrix with one column.}
}
\description{
The elements of the tangent space for a \eqn{p \in \mathbb{S}^{n}}
are \eqn{\mathcal{T}_p\mathbb{S}^n :=  \{v \in \mathbb{R}^n | \sum_i v_i p_i = 0\}}
}
